requirements.txt
streamlit
pandas
pdfplumber
Pillow
xlsxwriter
